Primer merge para el API

+++ database/migrations/2020_08_30_165002_create_personas_table.php
@@ -17,7 +17,7 @@ class CreatePersonasTable extends Migration
             $table->id();
             $table->timestamps();
             $table->string('nombre_completo')->unique();
-            $table->integer('porcentaje');
+            $table->double('porcentaje');
         });
     }
